requirements.txt
Adafruit-Blinka==8.50.0
adafruit-circuitpython-busdevice==5.2.10
adafruit-circuitpython-connectionmanager==3.1.2
adafruit-circuitpython-framebuf==1.6.6
adafruit-circuitpython-requests==4.1.8
adafruit-circuitpython-ssd1306==2.12.18
adafruit-circuitpython-typing==1.11.2
Adafruit-GPIO==1.0.3
Adafruit-PlatformDetect==3.76.1
Adafruit-PureIO==1.1.11
Adafruit-SSD1306==1.6.2
binho-host-adapter==0.1.6
blinker==1.9.0
board==1.0
cbor2==5.6.5
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorzero==2.0
docker==7.1.0
Flask==3.1.0
gpiozero==2.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
lgpio==0.2.2.0
luma==2.3.20
luma.core==2.4.2
luma.oled==3.14.0
MarkupSafe==3.0.2
pillow==11.1.0
pyarmor==9.0.7
pyarmor.cli.core==7.6.2
pycparser==2.21
pyftdi==0.56.0
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.0
pyusb==1.3.0
qrcode==8.0
requests==2.32.3
RPi.GPIO==0.7.1
rpi_ws281x==5.0.0
six==1.17.0
smbus==1.1.post2
smbus2==0.5.0
spidev==3.6
sysv-ipc==1.1.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
packaging==24.2
rpi-lgpio==0.6
pytz
